+++ platform/linux/include/config-linux.h
@@ -2,6 +2,6 @@
 
 #define DEFAULT_ROOT "./"
 
-#define REGISTRY_ROOT "./wapps"
+#define REGISTRY_ROOT "./registry"
 #define REGISTRY_EXT ".wapp"
 #define REGISTRY_VERSION_SEPARATOR ':'