+++ platform/nuttx/include/config-nuttx.h
@@ -2,7 +2,7 @@
 
 #define DEFAULT_ROOT "./"
 
-#define REGISTRY_ROOT "./wapps"
+#define REGISTRY_ROOT "./registry"
 #define REGISTRY_EXT ".wapp"
 #define REGISTRY_VERSION_SEPARATOR ':'